move stubbing modules into same directory.
