+++ test/stubbing/stubbing.rb
